requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aiohttp==3.11.18
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
coverage==7.8.0
cryptography==44.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
pytest-cov==6.1.1
watchfiles==1.0.5
websockets==15.0.1
python-gitlab==5.6.0
GitPython==3.1.44
PyGithub==2.6.1
requests==2.32.0
